Split common test from core tests
